Moved logic of search to SearchServices - Performing the search is now done in SearchServices instead of directly in the transcript controller.
Issue #155
